builder: provide purging support
Introduces the ability to purge a detected set of legacy pages from a
configured parent page. In the event a user configures the option
`confluence_purge` to "True", the builder will keep track of all known
descendants of the parent page. When the builder starts publishing
pages, any previously tracked descendant matching the identifier of a
recently uploaded page will be removed from the legacy list. During the
finalization phase, if any legacy pages still remain in the list, these
legacy pages will be removed.

Signed-off-by: James Knight <james.d.knight@live.com>