builder: provide header/footer support
Introduces the ability for a user to configure a header/footer file
which content can be to apply to all parsed paged. In the event a user
configures the option `confluence_header_file`, the writer will attempt
to read the contents of the file and inject it into each generated file
before the body text. Similar with `confluence_footer_file` which will
attempt to inject content after the body text.

An example use case is to add a "This page was been automatically
generated." message at the bottom of each published page.

Signed-off-by: James Knight <james.d.knight@live.com>