cleanup module version
The following commit cleans up sphinxcontrib-confluencebuilder's module
version (`__version__`) by:

 - Moving the reference into the `sphinxcontrib.confluencebuilder`
    namespace.
 - Marking an interim version for this development cycle.

The version value has been bumped from "0.4.0" to "0.6.0rc" for two
(2) reasons. First, the release of this module's v0.5 release did not
increment this version reference (which would have had a value of
"0.5.0"). Secondly, to indicate the next release cycle, the version is
bumped to "0.6.0" with a development cycle "rc".

This change also adjust bumpversion's setting file to support the
release-cycle post-fix. To move out of the release cycle phase, the part
value `release` can be used.

Signed-off-by: James Knight <james.d.knight@live.com>