publisher: rework publisher for rest and xml-rpc support
This is a rework of sphinxcontrib-confluencebuilder publisher to support
both Confluence REST API and continued support of XML-RPC API. This
change drops the dependency on the "confluence" PyPI package and
introduces a dependency for the "requests" PyPI package. Both API types
are now provided through an internal implementation based off of the
"requests" and "xmlrpclib" libraries.

By default, the REST API will be attempted first. If REST support is
not detected, the publisher will then check if the XML-RPC API is
available. In Confluence 5.5+ environments, it is expected the REST API
will always be used. For Confluence 4.0 to 5.4 environments, the REST
API will be attempted, fail and then XML-RPC API will be used (assuming
XML-RPC API is enabled by a Confluence administrator). If a user of this
modules wishes to explicitly disable a certain API type, they may do so
through either of the following configuration options:

    confluence_disable_rest = True
     (or)
    confluence_disable_xmlrpc = True

While these options are not necessarily needed, it is recommended for
Confluence 4.0 to 5.4 users to configure `confluence_disable_rest` to
prevent the initial REST call on start-up (to prevent an unneeded
request to the Confluence server).

This change now permits the ability to perform anonymous publishing (if
the Confluence server is configured to permit anonymous modifications).
By providing an empty `confluence_server_user` and
`confluence_server_pass` options, it is assumed the user wishes to
publish anonymously.

This change has other note-worthies:
 - Python 3.x series should work again.
 - Exception details are now more explicit.
 - Validates configure space name with the server on publish connection
    attempt.

Signed-off-by: James Knight <james.d.knight@live.com>