publisher: provide timeout support
The following allows a user to adjust the timeout value for an API call
using the `confluence_timeout` option. By default, the timeout calls
are based on the "requests" and "xmlrpclib" libraries (which is
approximately thirty (30) seconds). A user can provide a timeout value
(in seconds) to reduce the time before timeout.

Signed-off-by: James Knight <james.d.knight@live.com>