publisher: track descendants of space
When a user configures publishing to a space root (instead of a parent
page), this extension was not tracking descendants of the space. This
would prevent users from using the purge capability on a space
dedicated to published documentation. With this change, if no parent
page is provided, it will populate known descendants of the configured
space.

Signed-off-by: James Knight <james.d.knight@live.com>