rest: handle confluence returning empty json response
Confluence REST GET/POST/PULL request should return response data to
parse. There have been some observations where interacting with a
Confluence Cloud instance with improper credentials which resulted in a
HTTP 200 response with no response data.

The following adds a new exception type to handle the odd case of not
receiving response data. This makes the assumption that this error case
is a result of an authentication issue (client or server-based).
Ideally, no one will ever experience this issue.

Signed-off-by: James Knight <james.d.knight@live.com>