publisher: ensure used space name is uppercase
If the name of the target space does not match the space name casing on
the Confluence instance, REST API calls will not find the target space.
This has been observed on Confluence Cloud instances. This commit forces
the space name to be uppercase to mitigate this issue. No ill effects
observed when using XML-RPC API.

Signed-off-by: James Knight <james.d.knight@live.com>