publisher: handle unknown macro migration
When publishing content using the XML-RPC API, the process will convert
the Wiki format to a `storage` format using an API call. In the event
that the Confluence server cannot handle a macro translation, the call
will fail. In the event that a page cannot perform such a translation,
a warning message will be generated and the legacy page will remain
untouched (will not be updated or purged (if enabled)).

Signed-off-by: James Knight <james.d.knight@live.com>