publisher: handle publish user permission errors
If the configured Confluence user does not have permission to write or
remove pages, both REST and XML-RPC APIs indicate a permission failure.
This commit captures the permission issues when attempting to add or
remove a page, and reflects a detailed message to the user of the
permission issues for the current user.

Signed-off-by: James Knight <james.d.knight@live.com>