writer: allow anchor macro to be restricted
This is a compatibility support change. For Confluence versions older
then v5.5, the Anchor macro was not available as a Wiki markup. If a
page is submitted for conversion with a target node on these legacy
versions, the page will not convert (and cannot be published). For
situations where a user cannot alter a given document's content, this
commit provides a hidden option to allow a user to restrict the
generation of a Confluence macro. For the anchor macro, a user can
restrict its generation by using the following in their configuration
file:

    confluence_restricted_macros = ['anchor']

In the event that restricted macro is detected, a warning message will
appear and the content will either be rendered in a non-macro form or
removed entirely from generation.

Signed-off-by: James Knight <james.d.knight@live.com>