builder: remove trailing forward slash on server url
When a user configures this extension with a `confluence_server_url`
value, the desired format is without a trailing prefix. For example:

    (YAY) confluence_server_url = 'https://intranet-wiki.example.com'
    (NAY) confluence_server_url = 'https://intranet-wiki.example.com/'

It order to improve usability for a configuring user, we'll allow both
options (but still recommend the former in the example) by checking for
a trailing forward-slash and removing it.

Signed-off-by: James Knight <james.d.knight@live.com>