publisher,writer: support standard indentation
On of the most common markups used with Sphinx is the reStructuredText's
indentation support [1]; however, default usage with this extension does
not provide any indentation visualization in Confluence. Confluence's
Wiki Markup [2], which is currently the primary internal markup used to
prepare Confluence pages has no real way to support indentations
(without the use of some macros). Confluence's Storage Format does
provide a means to provide indentation; by defining a paragraph with a
margin value (which with Confluence Editor seems to use as well).

Since we still rely on the Wiki format, an experimental storage
modification will be performed. A "reserved" keyword will be used to
indicate an indentation level. During the publishing phase, after the
Wiki markup is converted into a storage format, the storage format will
be modified to replace the indented paragraphs with a respective
margin offset.

When test's on Confluence Cloud and Confluence Server shows ideal
results, since this is playing with Confluence Storage Format, I'm
currently considering this experimental. While the option is being
enabled by default (since I believe most users will want this), an
option `confluence_experimental_indentation` is currently being made
available to allow a user to explicitly disable this experimental
indentation support.

[1]: http://docutils.sourceforge.net/docs/ref/rst/restructuredtext.html#indentation
[2]: https://confluence.atlassian.com/doc/confluence-wiki-markup-251003035.html
[3]: https://confluence.atlassian.com/doc/confluence-storage-format-790796544.html

Signed-off-by: James Knight <james.d.knight@live.com>