writer: do no allow bracket characters
The following explicitly removes the bracket characters ('[' and ']')
from being used in generated documentation. While Confluence Cloud (as
of ~2017-04) appears to be able to handle these characters when
escaped (ex. '\['), some existing and legacy versions of Confluence
server have no support for handling these characters (escaped or not)
when using the Confluence Wiki format (the Wiki conversion will
generated bad storage format data).

In the event that this extension moves to only using a Confluence
storage format, these characters can be considered again.

Signed-off-by: James Knight <james.d.knight@live.com>