writer: improve table markup
The following minimizes the implementation and capabilities of this
extension's handling of the reStructuredText's table markup [1]. The
formation of a table no longer performs any elaborate width calculation;
just simple forming of table columns/rows. With this minimization, there
are some added benefits, such as header-less tables and (simple) multi-
line content.

This table support does what it can with the limitation of table
customization in Confluence's Wiki markup [2]. If Confluence's Storage
Format becomes the primary internal format of this extension, more
elaborate table structures/configurations can be made.

[1]: http://docutils.sourceforge.net/docs/ref/rst/restructuredtext.html#tables
[2]: https://confluence.atlassian.com/doc/confluence-wiki-markup-251003035.html#ConfluenceWikiMarkup-Tables

Signed-off-by: James Knight <james.d.knight@live.com>