writer: improve support for citations, definitions and footnotes
Improve the use of the reStructuredText's citations [1], definitions [2]
and footnotes [3]. The following provides a series of functional and
visual improvements to the mentioned markups types. Both citations and
footnotes are visualized in simple table. The first column's value
contains the citation label or footnote identifier (depending on the
markup used). The second column provides the description for the
respective citation/footnote. When referencing a citation/footnote,
either the citation label or footnote identifier is seen in superscript
form. The selection of the reference will jump (using an anchor) to the
respective citation/footnote row.

This change also includes a re-work on a definition visualization. The
common form is a definition term followed by a quote of of its
definition. Related definitions will stack (when using Sphinx's
glossary). Definition classifications will trail the respective
definition term.

[1]: http://docutils.sourceforge.net/docs/ref/rst/restructuredtext.html#citations
[2]: http://docutils.sourceforge.net/docs/ref/rst/restructuredtext.html#definition-lists
[3]: http://docutils.sourceforge.net/docs/ref/rst/restructuredtext.html#footnotes

Signed-off-by: James Knight <james.d.knight@live.com>