test_publisher: add new manual publisher test
The following introduces `test_publisher`; a manually invoked test
option which can be used to publish a "validation set" to a target
Confluence server. The indent of this test is to help perform manual
validation of the interaction between this extension and various
Confluence instances (running various versions).

The `test_builder` class (or future unit test classes) should be still
used to validate Sphinx markup to expected Confluence markup (pre-
publishing). This validation set is not intended to be documentation.
This extension's documentation is covered in "doc/", while Sphinx and
reStructuredText documentation is made available by the respective
third party sources.

Signed-off-by: James Knight <james.d.knight@live.com>