tox: force pip upgrade
The Travis CI environment prepares a test environment with the most
recent version of pip available (v9.0.1); however, using this
version results in the following error [1]:

    AttributeError: '_NamespacePath' object has no attribute 'sort'

Adjusting the tox requirements to forcefully update pip to the
newest version for the interim. When pip v9.0.2+ is available, this
commit can be reverted.

[1]: https://github.com/tonybaloney/sphinxcontrib-confluencebuilder/issues/45

Signed-off-by: James Knight <james.d.knight@live.com>