test: move build output checks into expected files
When testing generated output to expected output, move existing
expected output checks to an external file (see also [1]). This is
to not always rely on an available index when asserting line
indexes and improve validate of the entire output (ex. checking
for excess line separators).

This change cleans up the toctree validation so that the expected
output does not necessarily rely on the sibling tests. toctree test
now has its own child documents to use to validate a properly formed
table of contents. To reduce the following warning from other
documents being tested:

    WARNING: document isn't included in any toctree

Adding a series of orphan markups to hide the generated warnings.

[1]: b2b02ee615403d45d2f3350bc61796e84e5ef050

Signed-off-by: James Knight <james.d.knight@live.com>