builder: remove detected document titles from output
When generating documents for Confluence, the following adds the
ability (by default) to remove any detected document titles from
the output. Since document titles are managed outside the document's
content in Confluence, there's commonly no need to represent the
title twice for a page:

     <confluence page title>    <-- first title
     <confluence page meta>
     <confluence content>       <-- repeated contents title

In some cases, a builder may not always want to remove the title
value from contents (in cases when publishing with a page prefix
value). To disable this capability, the configuration option
"confluence_remove_title" can be set to "False".

Signed-off-by: James Knight <james.d.knight@live.com>