test: allow publish test to accept input
To help improve manual test publish scenarios, allowing the publish
test to be configurable by input when the "--input" argument is
provided. For example:

    tox -e publisher -- --input

This is to assist in quickly pushing to "devel" (and possibly other
test containers) with out having to always prepare proper
environment settings or manually adjusting the test instance.

Signed-off-by: James Knight <james.d.knight@live.com>