experimental: rework/minimize indentation support
The original experimental implementation to support multi-level
quote blocks in Confluence handled as follows:

 - Injected keywords into sections of data that should be indented.
 - Process document data via Confluence to convert to Wiki format.
 - Post-process the data by adjusting paragraph entries that should
    be indented with margin values.

The issue with this experimental code is that processing that data
through an HTMLParse entity caused already encoded HTML data to be
decoded again. When publishing, this would cause 400 errors such as:

    REQ: PUT
    RSP: 400
    URL: <url>/rest/api/
    API: content
    MSG: Error parsing xhtml: Unexpected close tag </p>; expected </less-than>.
     at [row,col {unknown-source}]: [5,69]

The experimental code, while still crude, has been re-worked to a
very simplified version. Text that should be quoted is injected with
a start-token and the target quote level, while the end of a text
block is injected with an end-token. Instead of trying to update
margin blocks on wrapped Wiki-supported HTML blocks, the process
will just do a simple replacement of tokens with a respective
paragraph start/end tag (with an appropriate margin set). This
works for output since HTML paragraph elements will implicitly close
other block-level elements -- crude, but it works. This change
removes the need for HTML parsing and leaves HTML encoded data in
encoded form (which can be published gracefully).

Signed-off-by: James Knight <james.d.knight@live.com>