rest: interpret received data as utf-8
The following corrects the interpretation of received REST data from
Confluence. Confluence returns API data with a content type
'application/json', which by RFC-4627 shall be UTF-based and by
default UTF-8 [1]. Atlassian documents that Confluence should always
be UTF-8 [2]. This commit adjusts all interpreted JSON response data
as UTF-8.

This resolves issues where a document line may indicate, for
example, "Let’s see this issue" and be published/rendered on
Confluence as "Letâ€™s see this issue". With the two-pass processing
for the Wiki translator (Extension -> Wiki -> Storage), the bytes
response provided by Confluence showed the following:

  PRE | Let\xe2\x80\x99s see this issue.
 POST | <p>Let\xc3\xa2\xe2\x82\xac\xe2\x84\xa2s see this issue.</p>

By explicitly indicating that the received data is UTF-8, the
response was handled as expected:

  PRE | Let\xe2\x80\x99s see this issue.
 POST | <p>Let\xe2\x80\x99s see this issue.</p>

In the scenario where a Confluence instance is configured outside of
a recommended UTF encoding (UTF-16, UTF-32, etc.) users may
experiencing encoding issues. Since it's believed that this is a
rare scenario, additional UTF encoding types are not going to be
attempted to be supported (due to lack of test environments). This
a user provides a scenario where this is an issue with some example
data/configuration information, this fix can be improved.

[1]: https://www.ietf.org/rfc/rfc4627.txt
[2]: https://confluence.atlassian.com/doc/configuring-character-encoding-174010.html

Signed-off-by: James Knight <james.d.knight@live.com>