builder: add max-depth support
Supports the ability for a user to configure a maximum depth for nested
pages (when using hierarchy mode). When enabled, pages that exceed a
configured depth will be embedded into their parent page. The
master_doc [1] is considered to be at depth zero. If a maximum depth is
configured to zero ('confluence_max_doc_depth=0'), all page content*
will be embedded into the single master_doc file (* with the exception
of orphan documents); this is similar beheviour to the 'singlehtml'
builder [2] in Sphinx proper. By default, no maximum depth is set.

To support a maximum depth for pages requires some manipulation of
document's doctrees and various stages of the Sphinx pipeline. A common
practice used inside some single-document Sphinx builders uses an
document assembly stage with the help from an inline utility method [3].
While this can partially work with this extension, there's two
limitation's this has trying to integrate with this extension:
- Multiple documents can be published with a maximum depth of zero.
  While with a single master document could only produce a single
  published document when the maximum depth is configured to zero, if a
  project contains any orphan pages, these documents are validate
  candidates for publishing.
- This extension relies on the 'prepare_writing' stage to detect/manage
  document title/reference information to match Confluence-generated
  pages/anchors. Moving the title/reference management during or after
  the document writing stages is too complex to manage.

A notable change to consider is the need for document's doctree caching.
Sphinx's pipeline will read/prepare a document's doctree during a
write phase. Since we have a need to manipulate individual document's
doctree's during the 'prepare_writing' stage, we want to cache these
results for be used by a writer. This is achieved by overriding/hooking
the environment's 'get_doctree' method to support caching.

[1]: http://www.sphinx-doc.org/en/stable/config.html#confval-master_doc
[2]: http://www.sphinx-doc.org/en/stable/builders.html#sphinx.builders.html.SingleFileHTMLBuilder
[3]: https://github.com/sphinx-doc/sphinx/blob/25d18e65174a70504042f34bc9b773dea6c5ee64/sphinx/util/nodes.py#L279

Co-authored-by: John Teasdale <teasdale@uber.com>
Signed-off-by: James Knight <james.d.knight@live.com>