wiki-translator: fallback to anchor for internal links
There are cases from internal processing of doctree's resolves some
anchor-based links as internal links (primarily from squashing document
files [1]). The following removes the warning message with an assumption
that the local anchor already exists in the document. Removing the
warning shouldn't be an issue since if the reference does not exist when
parsing the internal domain information, Sphinx already generates a
warning reference.

[1]: 9d0dda592a9aabc10634019eb9842eccf705a8ae

Signed-off-by: James Knight <james.d.knight@live.com>