tox: remove explicit pylint config
CI for this extension uses 'pylint' to perform some code quality
validation of the source. The original invoke of this module provide a
configuration file '/.pylintrc' which by default does not exist
(developers could just modify their stage to include it to perform
custom validations). As of pylint 1.8.0 [1], providing a non-existent
configuration file results in an error (thus, the test will always
fail). This commit removes the (by default) unused configuration file
from the tox run.

[1]: https://github.com/PyCQA/pylint/commit/b9061a0596457cf89ddd3dcf65c8aae7fb241e69

Signed-off-by: James Knight <james.d.knight@live.com>