builder: only extract title value from a leading title element
A document's title value, which is used as the target Confluence page
name, originally pulled from the first section's title element. While
documents should be leading with a title value, there's no real
requirement for this. An orphan page may lead with a non-title element
(such as a paragraph of text) before introducing any sections (if any).
A user in this case will lead to unexpected results when the
configuration option 'confluence_remove_title' is enabled (default
state), and a section/title value is removed from the middle of a
document.

Adjusting the title value extraction by only looking at the first
element's title element of a document (which should be a section's
title). If one is not found, it is assumed that the document has no
proper title set.

Signed-off-by: James Knight <james.d.knight@live.com>