builder: provided flexible url handling
This extension requires a user to provide the URL path of the Confluence
instances to interact with when publishing. The configuration handling
was designed to take the "base" portion of the URL and the extension
would later attached the REST or XML-RPC bind point later.

This commit provides more flexibly in the configured URL by
automatically detecting URLs, for example, with the API bind point set.
The normalizing method will remove and cleanup the final URL before its
use. In the event that this extension normalizes the URL, a warning
message will be generated to the user (to promote a proper configuration
value).

Signed-off-by: James Knight <james.d.knight@live.com>