rest: handle non-json responses
The typical response in an error case provided a message value
containing JSON formatted data. In the case where a server returns
non-JSON data (for example, an invalid server URL), an exception will
be thrown when preparing the message to display. The following adds
exception handling and informs the user that the returned content is not
JSON-based data.

Signed-off-by: James Knight <james.d.knight@live.com>