publisher: add compatibility support for descendants fetch
It has been observed on some Confluence Server instances that a fetch of
descendants using CQL 'api/content/search', specifically the 'ancestor'
field [1], the result does not provide all direct children and their
children. When comparing with a Confluence Cloud instance, the query
works as expected.

The following introduces a 'getDescendentsCompat' method which searches
for all descendants from a base page no matter if the Confluence
instance is cloud or server based. Ideally, this method can be removed
in time when the oldest in-service Confluence Server version properly
supports 'ancestor' searching. This has been observed in Confluence
Server v6.0.6 (supported versions before and after have not be
explicitly examined for this issue).

[1]: https://developer.atlassian.com/server/confluence/cql-field-reference/#ancestor

Signed-off-by: James Knight <james.d.knight@live.com>