Improved SSL error messages.
The SSLError is a subclass of ConnectionError so it needs
to come before it in the except block. Before this commit
the error message indicated a connection problem but if
the error is an SSL error we should not fall back to the
XML RPC interface. In addition, we give a more descriptive
message about the underlying problem.

Signed-off-by: Logan Jones <loganasherjones@gmail.com>