builder: incorrectly flagged legacy page(s) on empty descendants
When populating a list of documents to purge from a space, a list
of legacy pages is populated once after the first publish attempt
(to easily fetch the base page identifier). In the event which the
base page does not have any descendants (usually first-ever publish
attempt), an empty list is generated and set into
`self.legacy_pages`. When this scenario occurs, the next publish
page will re-populate the known legacy pages list which could result
in pages which have already been flagged as not-legacy as legacy
again. To address this, an explicit check is made so that only if
`self.legacy_pages` is `None` will a list of descendants be
populated.

Signed-off-by: James Knight <james.d.knight@live.com>