publisher: handle various space names in desc-check
When a REST-based publishing event checks for page descendants (when
purging legacy pages), a CQL search request can be made when
attempting to find descendants for a space. The provided space name
is not quoted, and causes the CQL expression to break with some
valid space name values (for example, personal spaces which are
prefixes with a tilde (~) character). Enclosing the space name with
quotes prevent an expression error.

Signed-off-by: James Knight <james.d.knight@live.com>