translator: introduce to-be-completed support
Re-introduced a series of to-be-completed (via docutils documentation)
translation calls. When running Sphinx/docutils on a document, a series
of translation calls are required or an unknown exception error is
thrown. For a series of known common calls, the methods either pass or
skip-the-node to prevent hard failures. In time, these methods may be
modified depending on their use cases.

Signed-off-by: James Knight <james.d.knight@live.com>