translator: warn about unknown code languages once
For a documentation set containing a series of literal blocks with
unsupported language types, multiple warnings can be generated
for the same language type indicating that the language is not
supported. To minimize the warning-spam to a user, track code
languages that have already notified of a warning so they only
appear once a session.

Signed-off-by: James Knight <james.d.knight@live.com>