translator: build links even if they cannot be served
In cases where a referenced document does not have a title, this
extension is unable to properly generate a link to the document
(since there is no document to link to). The result of not
generating a link can lead to a more undesired state. A missing link
entry will show no indication that a link should have been in a
specific location, leading to either incomplete sentence structures
(grammar) or poorly formed link list (e.x. toctree's with empty
entries). Instead, allow links to be generated -- but with a
localized href.

Signed-off-by: James Knight <james.d.knight@live.com>