translator: properly check internal flag for external reference
When compiling external references, the node was checked for the
existence of the 'internal' attribute to see if the processing node
was an external-based reference. In select cases, the 'internal'
attribute can exist for a reference but with a value of 'False',
thus still a possible external link. Adding a proper check for the
'internal' attribute.

Signed-off-by: James Knight <james.d.knight@live.com>