translator: handle relaxed local anchor references
It is proper when trying to build a reference to a local area to
use reStructuredText references. However, in some documentation
sets, a document builder may crudely point to reference by using a
anchor value of a pending-generated content (i.e. using an anchor
value for a heading which will match the value in a generated
document).

The following provides a quirk which attempts to detect an external
reference which only contains an anchor value. If so, the anchor
value is stripped from the URI and pushed into the internal
reference method for processing (which may or may not work depending
on how the documents formed).

Signed-off-by: James Knight <james.d.knight@live.com>