publisher: track descendants with a set
Adjusting the descendants-fetch to operate using sets. In some rare
cases, it has been observed that a Confluence instance may return
information which repeats a page identifier. This can result in
unexpected purging when a document is marked as "has been uploaded"
but another identifier's existence will still result in the document
in being purged; using sets prevents this from occurring.

This change also adds the respective documentation for the
descendants methods.

Signed-off-by: James Knight <james.d.knight@live.com>