publisher: increase descendants fetch limit
Define a higher limit when searching for descendants from a
Confluence instance. This is to help minimize multiple calls when
searching for descendants (since the default limit is 25 results).

While not all Confluence instances will return the total limit set
(as it may be system restricted); however, the implementation should
be able to properly handle only acquiring only a subset of results
requested.

Signed-off-by: James Knight <james.d.knight@live.com>