builder: improve document title fetching
The following attempts to improve the determined title value for a
document to be published. Originally, there was "smart" checks to help
try exact the leading section's title value from a document. It would
only attempt to consider the first node after skipping a series of
"no-content nodes" (e.x. comments, etc.). This was to prevent having a
bulk of content and then using the first section which exists near the
bottom of a document.

The complications from the above approach is that there needs to be a
explicit list of node types to verify as "no-content" before taking the
next "real" node to consider. While categories like `nodes.Invisible`
was tried, not all directive types are properly categorized (in the
simplest example, a user could be using custom directives and do not
really care of what categories they are in).

Instead of having a list of skippable leading nodes, just look for the
first section element. This should produce a most likely desired title
value for a document for most users.

Signed-off-by: James Knight <james.d.knight@live.com>