publisher: handle confluence transaction failure
Confluence Cloud may (rarely) fail to complete a content request with an
OptimisticLockException/StaleObjectStateException exception:

    com.atlassian.confluence.core.persistence.confluence.StaleObjectStateException: javax.persistence.OptimisticLockException: Batch update returned unexpected row count from update [0]; actual row count: 0; expected: 1

It is suspected that this is just an instance timing/processing issue.
If this is observed, wait a moment and retry the content request. If it
happens again, the put request will fail as it normally would. This
should assist in users when publishing large document sets (200+).

Signed-off-by: James Knight <james.d.knight@live.com>