builder: remove references pointing to removed titles
When the 'confluence_remove_title' option is enabled (default: yes), the
detected title element from a document is removed from the final output.
If the page deploys a table of contents directive, the generated table
will provide a reference to the leading title element (which now no
longer exists). Instead of providing a broken link in the table of
contents, remove the entry from the list.

Signed-off-by: James Knight <james.d.knight@live.com>