builder: allow user to request a password prompt
Introduces a new option 'confluence_ask_password' to allow a user to
request this extension to prompt for a password to be used for
authenticating with a publish event to the configured Confluence
instance.

This feature is to assist in user's who wish to invoke a publish request
on a documentation set without having to manually adjust the
documentation's configuration, setup external environment/files or
provide a clear password in an interactive session. With this feature,
user's can invoke a build with the following option to a password prompt
to appear:

   sphinx-build [options] -D confluence_ask_password=1 <srcdir> <outdir>

This commit explicitly leaves the 'getpass' prompt value as empty and
provides the prompt information using system write calls. This is to
provide more feedback for Cygwin/MinGW sessions (invoked without
'winpty') which cannot handle pulling a password from 'getpass' (instead
of the session blocking with no input provided).

Signed-off-by: James Knight <james.d.knight@live.com>