rest: output more detailed rest failures
When an error occurs interacting with a Confluence instance, some
advanced error cannot be completely interpreted from the provided
'message' content. In addition to the message, Confluence may populate
more notifications inside "'data': { 'errors': [...] }". To handle this
and possibly other odd scenarios, just dump the entire response data in
the event of an unknown error.

Signed-off-by: James Knight <james.d.knight@live.com>