translator: allow inline markup inside references
The original reference processing pulled the reference label information
from the node (or the children of the reference node) and built a link
before skipping the processing of the remainder of the node. This breaks
the ability to render inline content that can be wrapped around a
reference node. For example, the following reStructuredText markup
should allow the construction of a bolded link:

    .. _a link: https://example.com/

    .. |a link| replace:: **a bolded link**

    |a link|_

Adjusting reference processing to not invoke pruning exception (when
possible to render inline markup). Due to various reference types, it is
not ideal to populate the internal 'self.context' array. Instead, a new
interim context array 'self._reference_context' can be populated and
each entry will be pop'ed when the reference processing is completed
(this works for references since it is (_should_) not be possible to
have references inside other references.

In order to support links to internal documents with inline markup
support, switching from a plain-text-link-body to a link-body type tag.
This element supports plain or rich text content and the content should
already be escaped (via other internal node content producers).

This is primarily being introduced to support future image linking
support.

Signed-off-by: James Knight <james.d.knight@live.com>